requirements.txt
fastapi
uvicorn[standard]
pandas
serpapi
python-multipart
google-generativeai
google-search-results

